26 Agosto flows/progetto_crew_flows/pyproject.toml
[project]
name = "progetto_crew_flows"
version = "0.1.0"
description = "progetto_crew_flows using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.165.1,<1.0.0",
    "duckduckgo-search>=6.2.0",
    "faiss-cpu>=1.12.0",
    "certifi>=2024.0.0",
    "httpx>=0.25.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
kickoff = "progetto_crew_flows.main:kickoff"
run_crew = "progetto_crew_flows.main:kickoff"
plot = "progetto_crew_flows.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
